requirements.txt
asgiref==3.5.2
automium==0.2.6
automium-web==0.1.1
click==8.1.3
colorama==0.4.5
crispy-bootstrap5==0.6
decorator==5.1.1
django-ckeditor==6.4.2
django-crispy-forms==1.14.0
django-js-asset==2.0.0
django-taggit==3.0.0
Flask==2.1.3
importlib-metadata==4.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
legacy==0.1.6
MarkupSafe==2.1.1
self==2020.12.3
sqlparse==0.4.2
tzdata==2022.1
utils==1.0.1
Werkzeug==2.1.2
zipp==3.8.1
